+++ .vscode/tasks.json
@@ -1,6 +1,4 @@
 {
-  // See https://go.microsoft.com/fwlink/?LinkId=733558
-  // for the documentation about the tasks.json format
   "version": "2.0.0",
   "tasks": [
     {
@@ -12,8 +10,6 @@
         "kind": "build",
         "isDefault": true
       },
-      "problemMatcher": [],
-      "detail": "Runs a specific Flutter test file on the Chrome platform."
     },
     {
       "label": "Native tests",
@@ -24,8 +20,6 @@
         "kind": "build",
         "isDefault": true
       },
-      "problemMatcher": [],
-      "detail": "Runs a specific Flutter test file on the Chrome platform."
     },
     {
       "label": "Web tests",
@@ -41,8 +35,6 @@
         "kind": "build",
         "isDefault": true
       },
-      "problemMatcher": [],
-      "detail": "Runs a specific Flutter test file on the Chrome platform."
     }
   ]
 }
